fix: simplify and fix tracking event for opening unit modal in curriculum visualiser
